+++ database/migrations/2024_04_23_202938_create_discounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
             $table->float('percentage');
-            $table->float('disconted_price');
+            $table->decimal('disconted_price', 20);
             $table->date('start_date');
             $table->date('end_date');
             $table->float('duration');
